fix: rename hook and update imports
