+++ src/components/GoogleAdsense/__test__/adsense.test.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import GoogleAdsense from '..';
 import { render } from '@/test';
